extract responses
