+++ backend/src/core/response.py
@@ -1,0 +1,61 @@
+"""The FastAPI responses used in this app."""
+
+from __future__ import annotations
+
+import json
+from http import HTTPStatus
+from typing import TYPE_CHECKING, Any, Mapping, override
+
+from fastapi.responses import JSONResponse
+
+from .schemas.api import APISchema
+
+if TYPE_CHECKING:
+    from starlette.background import BackgroundTask
+
+    from .schemas.cache import CacheItemSchema
+
+
+class APIResponse(JSONResponse):
+    """Used to wrap all API responses in a common JSON interface.
+
+    See `core.schemas.APISchema`.
+    """
+
+    @override
+    def __init__(
+        self,
+        content: Any,
+        cache: list[CacheItemSchema] | None = None,
+        status_code: int = HTTPStatus.OK,
+        headers: Mapping[str, str] | None = None,
+        media_type: str | None = None,
+        background: BackgroundTask | None = None,
+    ) -> None:
+        error_occurred: bool = status_code not in range(
+            HTTPStatus.OK, HTTPStatus.MULTIPLE_CHOICES
+        )
+        cache = [] if cache is None else cache
+        super().__init__(
+            APISchema(
+                status_code=status_code,
+                error=content if error_occurred else None,
+                data=content if not error_occurred else None,
+                cache=cache,
+            ).model_dump(),
+            status_code,
+            headers,
+            media_type,
+            background,
+        )
+
+    @override
+    def render(self, content: Any) -> bytes:
+        return json.dumps(
+            content,
+            ensure_ascii=False,
+            allow_nan=False,
+            indent=None,
+            separators=(',', ':'),
+            default=(str),  # Serialize datetime objects
+        ).encode('utf-8')
